+++ src/main/java/com/github/siboxd/fatturapa/model/IdFiscaleType.java
@@ -1,4 +1,4 @@
-package com.github.siboxd.fatturapa.model.invoiceheader;
+package com.github.siboxd.fatturapa.model;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
