+++ src/main/java/com/github/siboxd/fatturapa/model/invoiceheader/buyer/CessionarioCommittenteType.java
@@ -1,6 +1,6 @@
 package com.github.siboxd.fatturapa.model.invoiceheader.buyer;
 
-import com.github.siboxd.fatturapa.model.invoiceheader.IndirizzoType;
+import com.github.siboxd.fatturapa.model.IndirizzoType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
